treewide: use name in define and eval lines
For consistency, use full name instead of $(PKG_NAME) in define and eval
lines for all packages.

I've seen reviews that asked to do this before, and I am asking the same
during reviews now. To avoid this in the future, fix this treewide so
when people use existing packages as example, we will not have to
request this change anymore.

This makes all packages consistent with both LEDE and OpenWrt base
repositories.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>

Conflicts:
	libs/musl-fts/Makefile
	net/adblock/Makefile
	net/ddns-scripts/Makefile
	net/vpnbypass/Makefile